Trabalho_1_Distribuidos/requirements.txt
numpy==2.4.4
opencv-python==4.13.0.92
PyAudio==0.2.14
pyzmq==27.1.0
